feat: Implement Jwt Authentication on Tasks Controller

+++ backend/src/tasks/tasks.service.ts
@@ -1,9 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException, UseGuards } from '@nestjs/common';
 import { CreateTaskDTO } from './dto/create-task.dto';
 import { UpdateTaskDTO } from './dto/update-task.dto';
 import { Task } from './entities/task.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { type JwtPayload } from '../auth/types/jwt-payload.type';
+import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
+import { AuthGuard } from '@nestjs/passport';
 
 @Injectable()
 export class TasksService {
@@ -16,8 +19,12 @@ export class TasksService {
     throw new NotFoundException('Task not found!');
   }
 
-  async create(dto: CreateTaskDTO) {
-    const newTask = this.tasksRepository.create(dto);
+  @UseGuards(AuthGuard)
+  async create(dto: CreateTaskDTO, payload: JwtPayload) {
+    const newTask = this.tasksRepository.create({
+      userId: payload.sub,
+      ...dto
+    });
 
     await this.tasksRepository.save(newTask);
 
@@ -38,7 +45,10 @@ export class TasksService {
     return task;
   }
 
-  async update(id: string, dto: UpdateTaskDTO) {
+  @UseGuards(AuthGuard)
+  async update(id: string, dto: UpdateTaskDTO, payload: JwtPayload) {
+    if (payload.sub !== id) throw new UnauthorizedException(`You can't change another user task.`);
+
     const updatedTask = await this.tasksRepository.preload({ id, ...dto });
 
     if (!updatedTask) this.throwNotFoundException();
@@ -46,7 +56,10 @@ export class TasksService {
     return await this.tasksRepository.save(updatedTask);
   }
 
-  async remove(id: string) {
+  @UseGuards(JwtAuthGuard)
+  async remove(id: string, payload: JwtPayload) {
+    if (payload.sub !== id) throw new UnauthorizedException(`You can't delete another user task.`);
+
     const task = await this.findOne(id);
 
     return await this.tasksRepository.remove(task);
